+++ R/tableby.control.R
@@ -142,11 +142,12 @@ tableby.control <- function(
 
 add_tbc_stats_labels <- function(x) {
   start <- list(
-    Nmiss="N-Miss", Nmiss2="N-Miss", meansd="Mean (SD)", meanse = "Mean (SE)", meanpmsd="Mean &pm; SD", meanpmse = "Mean &pm; SE", medianrange="Median (Range)",
+    Nmiss="N-Miss", Nmiss2="N-Miss", Nmisspct="N-Miss (%)", Nmisspct2="N-Miss (%)",
+    meansd="Mean (SD)", meanse = "Mean (SE)", meanpmsd="Mean &pm; SD", meanpmse = "Mean &pm; SE", medianrange="Median (Range)",
     median="Median", medianq1q3="Median (Q1, Q3)", q1q3="Q1, Q3", iqr = "IQR",
     mean = "Mean", sd = "SD", var = "Var", max = "Max", min = "Min", meanCI = "Mean (CI)", sum = "Sum",
     gmean = "Geom Mean", gsd = "Geom SD", gmeansd = "Geom Mean (Geom SD)", gmeanCI = "Geom Mean (CI)",
-    range="Range", Npct="N (Pct)", Nevents="Events", medSurv="Median Survival",
+    range="Range", Npct="N (%)", Nrowpct="N (%)", Nevents="Events", medSurv="Median Survival",
     medTime = "Median Follow-Up", medianmad="Median (MAD)", Nsigntest = "N (sign test)",
     overall = "Overall", total = "Total", difference = "Difference"
   )